refactor: AddExpenseCategoryModal 리팩토링
