+++ src/app/home/setting/expenseCategory/_components/AddExpenseCategoryModal.type.ts
@@ -1,0 +1,11 @@
+import { ExpenseCategoryAction } from "../page";
+
+export interface AddExpenseCategoryModalProps {
+    newExpenseCategoryInputRef: React.RefObject<HTMLInputElement>;
+    expenseCategoriesDispatch: React.Dispatch<ExpenseCategoryAction>;
+}
+
+export interface UseAddExpenseCategoryModalProps {
+    newExpenseCategoryInputRef: React.RefObject<HTMLInputElement>;
+    expenseCategoriesDispatch: React.Dispatch<ExpenseCategoryAction>;
+}
